# ENH improve init_root of HGBT TreeGrower