#### Reference Issues/PRs
None

#### What does this implement/fix? Explain your changes.
This PR improves the (compute/time) performance of HGBT a little bit by reducing a sum over n_samples in each new root node to a sum over n_bins.

#### Any other comments?
This also makes the Cython function `sum_parralel` obsolete.