@lorentzenchr 2025-02-22
> Do you have a quick benchmark to show the slight improvement?

Some very rough benchmark numbers running `python benchmarks/bench_hist_gradient_boosting_higgsboson.py --n-trees 100`

#### main branch
```
Training set with 8800000 records with 28 features.
Fitting a sklearn model...
Binning 1.971 GB of training data: 4.917 s
Fitting gradient boosted rounds:
Fit 100 trees in 62.501 s, (3100 total leaves)
Time spent computing histograms: 36.928s
Time spent finding best splits:  0.207s
Time spent applying splits:      8.364s
Time spent predicting:           1.833s
fitted in 62.696s
predicted in 7.975s, ROC AUC: 0.8228, ACC: 0.7415
```

#### PR branch
```
Training set with 8800000 records with 28 features.
Fitting a sklearn model...
Binning 1.971 GB of training data: 4.751 s
Fitting gradient boosted rounds:
Fit 100 trees in 60.776 s, (3100 total leaves)
Time spent computing histograms: 36.435s
Time spent finding best splits:  0.204s
Time spent applying splits:      8.070s
Time spent predicting:           1.888s
fitted in 60.970s
predicted in 8.478s, ROC AUC: 0.8228, ACC: 0.7415
```

While the difference is within the variation, running several times confirms the slight difference.